fix(audit): 19 critical bugs — unreachable code, security, datetime compare, etc
- nanny.py: who_check tool was nested inside elif tuple, never reachable
- devops.py:1178: read_file base path was /home/user/many (typo) → /home/user/family-hq
- devops.py: FamilyMode toggle preserves started_at (was wiped on each call)
- devops.py: _normalize_rule_dict recurses into and/or sub-rules
- devops.py: notebook_sync_rules compares rules tab (was tasks tab, always 'skipped')
- models.py: Subscription.amount, UtilityBill.amount → Float (was Integer)
- access_control.py: get_other_owner guards non-owner callers
- dispatcher.py: remove dead 'or not data.get(agents)' branch
- scheduler/digest.py: deleted (dead module, superseded by morning_brief)
- healthcheck.py: external bots now upsert (was OR IGNORE → never updated)
- reminders.py: narrow 24h window to <1h so each event fires once not twice
- automation.py: narrow weather eval except; cache TuyaClient; alert_ended uses parsed datetime not lexical string compare
- tuya.py: shared aiohttp session (was leaking sockets); fix looks_offline always-True OR
- transcribe.py: aiofiles instead of sync open() in async function
- claude_client.py: parenthesise 'raise ... from gemini_err or claude_err'

+++ src/db/models.py
@@ -244,7 +244,7 @@ class Subscription(Base):
     __tablename__ = "subscriptions"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String, nullable=False)
-    amount: Mapped[float] = mapped_column(Integer, nullable=False)  # SQLite REAL
+    amount: Mapped[float] = mapped_column(Float, nullable=False)
     currency: Mapped[str] = mapped_column(String, nullable=False, default="UAH")
     billing_day: Mapped[int] = mapped_column(Integer, nullable=False, default=1)  # 1-28
     next_charge: Mapped[str | None] = mapped_column(String, nullable=True)
@@ -256,7 +256,7 @@ class UtilityBill(Base):
     __tablename__ = "utility_bills"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     kind: Mapped[str] = mapped_column(String, nullable=False)  # газ/свет/вода/интернет/квартплата
-    amount: Mapped[float] = mapped_column(Integer, nullable=False)
+    amount: Mapped[float] = mapped_column(Float, nullable=False)
     currency: Mapped[str] = mapped_column(String, nullable=False, default="UAH")
     paid_at: Mapped[str | None] = mapped_column(String, nullable=True)
     due_at: Mapped[str | None] = mapped_column(String, nullable=True)
